Update UID_ISO_FIPS_LookUp_Table.csv
The population of "New York City, New York, US" is off by 2.7 million.  The value should include all five boroughs.  The number in the file was 5,803,210.  Per a 2016 U.S. Census Bureau report, the value is 8,537,673.  I suspect that someone transposed the first two digits in your CSV file, so I am proposing the change to 8,503,210.