Updates to France 14/4 to present
We have identified that our logic for France, France was double-counting cases from the overseas regions and departments. These cases are included in the total reported by Sante Publique, but we have those regions (Martinique, French Guiana, Mayotte, La Reunion, Guadeloupe) as separate entries that collectively contribute to the total cases in France. We are adjusting the time series data to correct this issue.